requirements.txt
# Telemetry & API Polling
requests>=2.31.0

# ML & RL
torch>=2.0.0
transformers>=4.35.0
gymnasium>=0.29.0
stable-baselines3>=2.2.0
scikit-learn>=1.3.0

# Dashboard & Visualization
streamlit>=1.28.0
prometheus-client>=0.18.0
matplotlib>=3.8.0
seaborn>=0.13.0

# Utilities
pandas>=2.1.0
numpy>=1.24.0
python-dotenv>=1.0.0

# Dev
pytest>=7.4.0
black>=23.10.0
flake8>=6.1.0
